+++ app/assets/stylesheets/application.scss
@@ -58,14 +58,6 @@ a {
 	}
 }
 
-.visible {
-	display: block;
-}
 .disvisible {
 	display: none;
 }
-button                  .show-when-disabled { display: none; }
-button[disabled]        .show-when-disabled { display: initial; }
-
-button                  .show-when-enabled { display: initial; }
-button[disabled]        .show-when-enabled { display: none; }
