+++ app/javascript/controllers/alert_controller.js
@@ -10,3 +10,22 @@ export default class extends Controller {
 		}, 100); // 0.1秒後にフォームを実際に送信
 	}
 }
+
+document.getElementById('js-active-new').addEventListener('click', function() {
+	        document.getElementById('js-active-edit').classList.remove('active');
+	        document.getElementById('js-active-reiatu').classList.remove('active');
+	        this.classList.add('active');
+});
+
+document.getElementById('js-active-edit').addEventListener('click', function() {
+	        document.getElementById('js-active-new').classList.remove('active');
+	        document.getElementById('js-active-reiatu').classList.remove('active');
+	        this.classList.add('active');
+});
+
+document.getElementById('js-active-reiatu').addEventListener('click', function() {
+	        document.getElementById('js-active-new').classList.remove('active');
+	        document.getElementById('js-active-edit').classList.remove('active');
+	        this.classList.add('active');
+});
+
