+++ app/views/posts/_before_login_reiatus_button.html.erb
@@ -1,8 +1,10 @@
-<div class="me-4 before-button" style="color: #026CE8; text-decoration:none; opacity: 0.8;">
-  <i class="fa fa-thumbs-o-up me-1" aria-hidden="true"></i>霊圧
-  <% if post.reiatus.count != 0 %>
-    <span style="color: #026CE8;">
-      <%= post.reiatus.count %>
-    </span>
-  <% end %>
+<div data-controller="button">
+  <div class="me-4 before-button" style="color: #026CE8; text-decoration:none; opacity: 0.8;" data-action="click->button#before_reiatu_button">
+    <i class="fa fa-thumbs-o-up me-1" aria-hidden="true"></i>霊圧
+    <% if post.reiatus.count != 0 %>
+      <span style="color: #026CE8;">
+        <%= post.reiatus.count %>
+      </span>
+    <% end %>
+  </div>
 </div>
