+++ app/views/profiles/show.html.erb
@@ -17,10 +17,10 @@
           <td><%= image_tag current_user.avatar.url, class: 'rounded-circle', size: '50x50', alt: "アバター" %></td> 
         </tr>
       </table>
-      <div class="btn-group mb-3" role="group" aria-label="Basic outlined example">
-        <%= link_to t('.my_post'), my_post_order_profiles_path, class: "active btn btn-outline-dark", id: "js-active-new", data: { turbo: true, turbo_method: :get } %>
-        <%= link_to '霊圧', my_reiatu_order_profiles_path, class: "btn btn-outline-dark", id: "js-active-reiatu", data: { turbo: true, turbo_method: :get } %>
-        <%= link_to t('.draft'), my_draft_order_profiles_path, class: "btn btn-outline-dark", id: "js-active-edit", data: { turbo: true, turbo_method: :get } %>
+      <div class="btn-group mb-3" role="group" aria-label="Basic outlined example" data-controller="button">
+        <%= link_to t('.my_post'), my_post_order_profiles_path, class: "active btn btn-outline-dark", data: { turbo: true, button_target: 'new', action: 'click->button#new_button' } %>
+        <%= link_to '霊圧', my_reiatu_order_profiles_path, class: "btn btn-outline-dark", data: { turbo: true, button_target: 'edit', action: 'click->button#edit_button' } %>
+        <%= link_to t('.draft'), my_draft_order_profiles_path, class: "btn btn-outline-dark", data: { turbo: true, button_target: 'reiatu', action: 'click->button#reiatu_button' } %>
       </div>
       <div class="col-12">
         <%= render 'profiles/profile_order', posts: @myposts %>
